trunk: Fix conversion and unused variable warnings in OperationMultiEvalStreaming_multImpl and OperationMultiEvalModMaskStreaming. [datadriven]
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@5369 4eea3252-f0fb-4393-894d-40516dce545b

+++ datadriven/src/sgpp/datadriven/operation/hash/OperationMultiEvalModMaskStreaming/OperationMultiEvalModMaskStreaming.cpp
@@ -174,8 +174,8 @@ void OperationMultiEvalModMaskStreaming::prepare() {
 void OperationMultiEvalModMaskStreaming::recalculateLevelIndexMask() {
 
     //TODO: does the padding work? test
-//    uint32_t localWorkSize = 24;
-    uint32_t localWorkSize = this->getChunkGridPoints();
+//    size_t localWorkSize = 24;
+    size_t localWorkSize = this->getChunkGridPoints();
 
     size_t remainder = this->storage->size() % localWorkSize;
     size_t padding = 0;
